+++ src/test/java/org/example/cucumber/StepsSearchItems.java
@@ -1,6 +1,5 @@
 package org.example.cucumber;
 
-import io.cucumber.java.ru.Допустим;
 import io.cucumber.java.ru.И;
 import io.cucumber.java.ru.Когда;
 import io.cucumber.java.ru.Тогда;
@@ -13,23 +12,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static com.codeborne.selenide.Selenide.$;
-import static com.codeborne.selenide.Selenide.open;
 import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class StepsSearchItems {
 
     private static final Logger log = LoggerFactory.getLogger(StepsSearchItems.class);
-
-    @Допустим("^открыта домашняя страница \"([^\"]*)\"$")
-    public void openPage(String pageUrl) {
-        open(pageUrl);
-        getWebDriver().manage().window().maximize();
-        String currentPageUrl = getWebDriver().getCurrentUrl();
-        assertEquals(pageUrl, currentPageUrl, "Адреса не совпадают");
-    }
 
     @Когда("в поле поиска введено значение {string}")
     public void enterValue(String value) {
