+++ src/test/resources/features/searchItemsTest.feature
@@ -5,7 +5,7 @@
 Функционал: Поиск товара
 
   Предыстория:
-    Допустим открыта домашняя страница "https://demowebshop.tricentis.com/"
+    Допустим открыта страница "https://demowebshop.tricentis.com/"
 
   Сценарий: Поиск ноутбуков
     Когда в поле поиска введено значение "laptop"
